+++ kaithem/src/chandler/blendmodes.py
@@ -1,4 +1,3 @@
-
 import numpy
 import time
 import random
@@ -7,6 +6,7 @@ import weakref
 
 from ..kaithemobj import kaithem
 from . import core
+from . import universes
 
 
 def getUniverse(u):
@@ -16,12 +16,6 @@ def getUniverse(u):
     except KeyError:
         oldUniverseObj = None
     return oldUniverseObj
-
-
-if not hasattr(core, "blendmodes"):
-    blendmodes = weakref.WeakValueDictionary()
-else:
-    blendmodes = core.blendmodes
 
 
 def getblenddesc(mode):
@@ -36,18 +30,20 @@ def getblenddesc(mode):
 
     try:
         return blendmodes[mode].description
-    except:
+    except Exception:
         try:
             return blendmodes[mode].__doc__
-        except:
+        except Exception:
             return ""
 
 
-class BlendMode():
+class BlendMode:
     autoStop = True
 
 
-core.getblenddesc = getblenddesc
+blendmodes: weakref.WeakValueDictionary[str, BlendMode] = weakref.WeakValueDictionary()
+
+
 core.BlendMode = BlendMode
 core.blendmodes = blendmodes
 
@@ -59,13 +55,18 @@ def makeBlankArray(l, v=0):
 
 class flicker_blendmode(BlendMode):
     "Blend mode based on physical model flickering"
-    parameters = {"gustiness": ("Gustiness", "number", "", 0.2),
-                  "lowpass": ("Lowpass", "number", "", 0.06),
-                  "topple_chance": ("Windiness", "number", "", 0.1),
-                  "agility": ("Flame agility", "number", "", 0.15),
-                  "group": ("Group", "number", "Groups of this many channels flicker together, e.g. to make RGB look right", 3),
-
-                  }
+    parameters = {
+        "gustiness": ("Gustiness", "number", "", 0.2),
+        "lowpass": ("Lowpass", "number", "", 0.06),
+        "topple_chance": ("Windiness", "number", "", 0.1),
+        "agility": ("Flame agility", "number", "", 0.15),
+        "group": (
+            "Group",
+            "number",
+            "Groups of this many channels flicker together, e.g. to make RGB look right",
+            3,
+        ),
+    }
     description = """Effects layer that simulate a natural flame-like flickering effect. Channel value determines how much flicker to apply to that channel.
 To use this, add a layer below containing the base colors. The flicker layer will randomly darken them according to it's simulation algorithm."""
 
@@ -109,10 +110,10 @@ To use this, add a layer below containing the base colors. The flicker layer wil
         lp = t60 * 0.05
         self.wind = 1 * lp + self.wind * (1 - lp)
 
-        if random.random() < self.scene.blendArgs['gustiness'] * t60:
+        if random.random() < self.scene.blendArgs["gustiness"] * t60:
             self.wind = max(random.normalvariate(1.3, 1), 1.2)
         if random.random() < 0.08:
-            rr = self.scene.blendArgs['agility']
+            rr = self.scene.blendArgs["agility"]
             self.riserate = random.normalvariate(rr, rr / (4.0))
 
         # Get the per-universe time interval
@@ -120,8 +121,8 @@ To use this, add a layer below containing the base colors. The flicker layer wil
         self.last_per[u] = time.time()
 
         ctr = 0
-        tc = self.scene.blendArgs['topple_chance']
-        lps = self.scene.blendArgs['lowpass']
+        tc = self.scene.blendArgs["topple_chance"]
+        lps = self.scene.blendArgs["lowpass"]
 
         # This algorithm is pretty tricky and I'm not sure how to properly implement it in numpy.
         # So we're doing it one pixel at a time in python
@@ -130,12 +131,12 @@ To use this, add a layer below containing the base colors. The flicker layer wil
         heights = self.heights[u]
         heights_lp = self.heights_lp[u]
 
-        group = int(self.scene.blendArgs['group'])
+        group = int(self.scene.blendArgs["group"])
         for k in numpy.nonzero(values)[0]:
             k = int(k)
             # Detect RGB groups of N, put them all together.
             # Reset group on finding a gap to account for typical DMX layouts
-            if ((not (ctr % group)) or k - lastk > 1):
+            if (not (ctr % group)) or k - lastk > 1:
                 t = random.random()
                 ctr = 0
                 nv = random.triangular(0, 1, 0.35)
@@ -146,8 +147,7 @@ To use this, add a layer below containing the base colors. The flicker layer wil
 
             lastk = k
             if t < (tc * self.wind * t60):
-                heights[k] = min(
-                    1 - (nv * (values[k] / 255.0)), heights[k] + 0.1)
+                heights[k] = min(1 - (nv * (values[k] / 255.0)), heights[k] + 0.1)
             else:
                 if heights[k] < 1 and values[k] > 0:
                     heights[k] += rise * t60
@@ -157,21 +157,36 @@ To use this, add a layer below containing the base colors. The flicker layer wil
             f = min(1, lp * t60)
             heights_lp[k] = heights_lp[k] * (1 - f) + heights[k] * (f)
 
-        old *= (alphas * alpha * numpy.minimum(heights_lp, 1)) + \
-            1 - (alpha * alphas)
+        old *= (alphas * alpha * numpy.minimum(heights_lp, 1)) + 1 - (alpha * alphas)
         return old
 
 
-blendmodes['flicker'] = flicker_blendmode
+blendmodes["flicker"] = flicker_blendmode
 
 
 class vary_blendmode_np(BlendMode):
     "Ads random variation, basically a random time varying gel"
-    parameters = {"interval": ("Change Interval", "number", "How many seconds between changes", 1.2),
-                  "rinterval": ("Randomize Interval", "number", "Amount to randmoly vary change interval", 0.5),
-                  "speed": ("Speed", "number", "How fast to change", 0.015),
-                  "mode": ("Mode", "number", "Mode value for the triangular distribution that defines the random values", 0.8),
-                  }
+    parameters = {
+        "interval": (
+            "Change Interval",
+            "number",
+            "How many seconds between changes",
+            1.2,
+        ),
+        "rinterval": (
+            "Randomize Interval",
+            "number",
+            "Amount to randmoly vary change interval",
+            0.5,
+        ),
+        "speed": ("Speed", "number", "How fast to change", 0.015),
+        "mode": (
+            "Mode",
+            "number",
+            "Mode value for the triangular distribution that defines the random values",
+            0.8,
+        ),
+    }
 
     def __init__(self, scene):
         self.vals = {}
@@ -201,29 +216,29 @@ class vary_blendmode_np(BlendMode):
         # It's always changing so always rerender
         self.scene.rerender = True
         if time.time() > self.ntt:
-            interval = self.scene.blendArgs['interval']
-            rnd = self.scene.blendArgs['rinterval']
-            avg = self.scene.blendArgs['mode']
-            nv = numpy.random.triangular(
-                0, max(min(1, 1 - avg), 0), 1, values.shape)
+            interval = self.scene.blendArgs["interval"]
+            rnd = self.scene.blendArgs["rinterval"]
+            avg = self.scene.blendArgs["mode"]
+            nv = numpy.random.triangular(0, max(min(1, 1 - avg), 0), 1, values.shape)
             self.vals[u] = 1 - (nv * (values / 255.0))
-            self.ntt = time.time() + random.triangular(interval -
-                                                           rnd, interval + rnd, interval)
+            self.ntt = time.time() + random.triangular(
+                interval - rnd, interval + rnd, interval
+            )
 
-        lp = t60 * self.scene.blendArgs['speed']
+        lp = t60 * self.scene.blendArgs["speed"]
         if uobj:
             if not uobj.localFading:
                 lp = 1
-                uobj.interpolationTime = (
-                    1 / 60) / self.scene.blendArgs['speed']
+                uobj.interpolationTime = (1 / 60) / self.scene.blendArgs["speed"]
 
         self.vals_lp[u] = self.vals_lp[u] * (1 - lp) + self.vals[u] * lp
-        old *= numpy.minimum((self.scene.alpha *
-                             self.vals_lp[u]) + 1 - self.scene.alpha, 255)
+        old *= numpy.minimum(
+            (self.scene.alpha * self.vals_lp[u]) + 1 - self.scene.alpha, 255
+        )
         return old
 
 
-blendmodes['vary'] = vary_blendmode_np
+blendmodes["vary"] = vary_blendmode_np
 
 
 class exp_blendmode_np(BlendMode):
@@ -237,10 +252,12 @@ class exp_blendmode_np(BlendMode):
         self.last = time.time()
 
     def frame(self, u, below, values, alphas, alpha):
-        return (((below**(values / 100.0)) / 255.0**(values / 100.0)) * 255) * (alphas * alpha) + below * (1 - alphas * alpha)
+        return (((below ** (values / 100.0)) / 255.0 ** (values / 100.0)) * 255) * (
+            alphas * alpha
+        ) + below * (1 - alphas * alpha)
 
 
-blendmodes['gamma'] = exp_blendmode_np
+blendmodes["gamma"] = exp_blendmode_np
 
 
 class sparks_blendmode(BlendMode):
@@ -248,9 +265,18 @@ class sparks_blendmode(BlendMode):
     parameters = {
         "fadetime": ("Fade Time", "number", "How fast to fade out again.", 1),
         "interval": ("Interval", "number", "How often to do a spark", 3),
-        "variation": ("Variation", "number", "How much to vary the spark intensity", 0.3),
-        "group": ("Group", "number", "Detect groups of channels to connect, e.g. to make RGB look right", 3),
-
+        "variation": (
+            "Variation",
+            "number",
+            "How much to vary the spark intensity",
+            0.3,
+        ),
+        "group": (
+            "Group",
+            "number",
+            "Detect groups of channels to connect, e.g. to make RGB look right",
+            3,
+        ),
     }
 
     def __init__(self, scene):
@@ -277,18 +303,16 @@ class sparks_blendmode(BlendMode):
         lastk = ctr = 0
         x = False
         t = time.time()
-        group = self.scene.blendArgs['group']
+        group = self.scene.blendArgs["group"]
         # Groups of 3 are supposed to come on at the same time because of RGB triples.
         for k in numpy.nonzero(alphas)[0]:
-            if ((not (ctr % group)) or k - lastk > 1):
+            if (not (ctr % group)) or k - lastk > 1:
                 ctr = 0
-                nv = random.triangular(
-                    1 - self.scene.blendArgs['variation'], 1)
+                nv = random.triangular(1 - self.scene.blendArgs["variation"], 1)
                 if t > self.sparktimes.get(k, 0):
-                    c = self.scene.blendArgs['interval']
+                    c = self.scene.blendArgs["interval"]
                     x = True
-                    self.sparktimes[k] = t + \
-                        random.uniform(max(c / 3.0, 0.35), c * 2)
+                    self.sparktimes[k] = t + random.uniform(max(c / 3.0, 0.35), c * 2)
                 else:
                     x = False
             if x:
@@ -297,17 +321,19 @@ class sparks_blendmode(BlendMode):
             lastk = k
 
         # Supposed to be a first order lowpass filter to handle the fade out
-        t = (time.time() - self.last_per[u])
+        t = time.time() - self.last_per[u]
         self.last = time.time()
         # Calculate the decay constant
-        k = 1 / self.scene.blendArgs['fadetime']
-        y = math.e**-(k * t)
+        k = 1 / self.scene.blendArgs["fadetime"]
+        y = math.e ** -(k * t)
         # Exponential decay equation.
         self.vals_lp[u] *= y
 
         # The vals_lp are actually alphas that spark up and then fade out and control how much of the scene shows up
         # in that channel
-        return values * (self.vals_lp[u] * alpha * alphas) + old * (1 - (self.vals_lp[u] * alpha * alphas))
+        return values * (self.vals_lp[u] * alpha * alphas) + old * (
+            1 - (self.vals_lp[u] * alpha * alphas)
+        )
 
 
-blendmodes['sparks'] = sparks_blendmode
+blendmodes["sparks"] = sparks_blendmode